nodejs: Terminate request on error
