+++ minireq-node/test/common.ts
@@ -1,7 +1,11 @@
 import {
     makeSimpleTests,
     makeHeaderTests,
-    makeQueryTests
+    makeQueryTests,
+    makeResponseTypeTests,
+    makeSendTests,
+    makeSerializationTests,
+    makeStreamTests
 } from '../../test/index';
 
 import { makeRequest } from '../src/index';
@@ -11,7 +15,7 @@ const request = makeRequest();
 makeSimpleTests(request);
 makeHeaderTests(request);
 makeQueryTests(request);
-//makeResponseTypeTests(request);
-//makeSendTests(request);
-//makeSerializationTests(makeRequest);
-//makeStreamTests(request);
+makeResponseTypeTests(request);
+makeSendTests(request);
+makeSerializationTests(makeRequest);
+//makeStreamTests(request); // TODO: Get working
